Reorganize command into single file per best practices
