+++ go.mod
@@ -2,8 +2,9 @@ module sw-cli
 
 go 1.19
 
+require github.com/spf13/cobra v1.6.1
+
 require (
 	github.com/inconshreveable/mousetrap v1.0.1 // indirect
-	github.com/spf13/cobra v1.6.1 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 )
